Make drake_ros package to reuse systems
