+++ drake_ros/setup.py
@@ -1,0 +1,15 @@
+import setuptools
+
+setuptools.setup(
+    name='drake_ros',
+    version='0.1.0',
+    author='Shane Loretz',
+    author_email='sloretz@openrobotics.org',
+    description='Tools for interacting with Drake Simulator and ROS',
+    # long_description
+    # long_description_content_type
+    packages=setuptools.find_packages(),
+    # classifiers
+    # python_requires
+)
+
